+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the Imperative package will be documented in this file.
 
+## Recent Changes
+- Enhancement: Enable command arguments to change `{$Prefix}_EDITOR`. Added text coloration for terminal prompts. Changes made to support recent zowe cli work: 
+[zowe-cli#1672](https://github.com/zowe/zowe-cli/pull/1672)
+
 ## `5.11.1`
 
 - BugFix: Fixed the `login` and `logout` handlers, fixing the `li` and `lo` aliases.